// src/providers/upnance/__generated__/upnance/model/address-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Upnance
 * Upnance API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Country } from './country';

/**
 * 
 * @export
 * @interface AddressModel
 */
export interface AddressModel {
    /**
     * 
     * @type {string}
     * @memberof AddressModel
     */
    'addressLine1': string;
    /**
     * 
     * @type {string}
     * @memberof AddressModel
     */
    'addressLine2'?: string;
    /**
     * 
     * @type {string}
     * @memberof AddressModel
     */
    'postalCode': string;
    /**
     * 
     * @type {string}
     * @memberof AddressModel
     */
    'city': string;
    /**
     * 
     * @type {string}
     * @memberof AddressModel
     */
    'stateOrCounty': string;
    /**
     * 
     * @type {Country}
     * @memberof AddressModel
     */
    'country': Country;
}